MGPG-135 - Support Overriding / Enhance the signer in AbstractGpgMojo
Currently AbstractGpgMojo has a fixed set of supported signers (and how
they are constructed), it would be good to allow extensions (e.g. Tycho
is also using AbstractGpgMojo) to possibly override that aspect.

This extracts the creation of the signer into a (protected) method so
different types or alternative implementations can be used.